+++ explorer/client/src/utils/mockUtil.ts
@@ -1,6 +1,3 @@
-// Copyright (c) 2022, Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
 import mockTransactionData from './mock_data.json';
 
 export const findDataFromID = (targetID: string | undefined, state: any) =>
